backend/requirements.txt
Django==2.*
django-allauth
pylint
pylint-django
djangorestframework
